Set activerecord as a dep without installing it
Previously, the storeconfigs class was installing
version 2.2.2 of the active records gem.

This created an unneccary coupling with that
specific version of the gem.

It is possible that users may want to specify their
own activerecord package.

This commit removes the activerecord package declaration
and just sets it as a class requirement.