add anchors to puppet/master
As much as the class containment issues bother
me in Puppet, and despite the fact I resent
anchors, they are a fact of life, and are
occasionally needed.

This commit adds anchors to the classes
contained in the puppet::master class so that
I can specify dependencies on that class.