update puppet cert call
Previously, the call to puppet cert generate was
returning 24 instead of 0 and indicating failures
even when it was successful.

This commit updates the call to instead use

* puppet certificate generate
* puppet cert sign (puppet certificate sign seemed to be broken)
* and puppet certificate fin

instead of the single call.

I know it sucks to have to make 3 calls to replace 1,
but in my testing on Puppet 2.7.19, that 1 call just did not
work and these 3 calls did.