Add vhost file definitions
With the addition of the RedHat code elements and specifically
the inclusion of the apache service in the init, it is required
to explicitly call out the required vhosts that are expected to
be running. This is a Debian/Ubuntu based issue because RedHat
deployemnts do not use the passenger/apache2 version of the
puppetmaster code.